redaction of explore page bands CSS changes
sorry Gus! All CSS sizing needs to be relative to the rem for the site instead of by px or vw. I should put that in the README. That'll be the next commit. Your changes did make me realize that the mobile font size for the band page was too large and so it is reduced. I've gotten good feedback on the tiling of the band explore so for now we will keep it that way. I'd like all the band names be displayed in full without having to ellipse them and tiling is the only way I know to do that without making things wonky (unless we get some javascript font sizing based on character length stuff going on).